openall now works with matlab 2017b
have not tested with matlab 2017a. But in 2017b at least the return array of dir is different. it used to be row 4 was a boolean meaning isdir. now row 4 is 0 if it IS a directory and some big old number if it isn't. row 5 is the isdir boolean now. this essentially REVERSES the behavior of existing code. fixed here for matlab 2017b because that is what i'm using but holy shit.